BugID(s):  VOIP-64473
Convert .c files to .cpp and remove extern "C" statements
Fix the few issues caused by C++ rather than C compiler rules


git-svn-id: https://subversion.polycom.com/SVN/RepoMisc/TestUtilities/trunk/SIPped/SIPped/src@26767 85c8ee13-dc6b-4ec3-b49e-93a6737f2991

+++ send_packets.cpp
@@ -51,6 +51,7 @@
 #ifndef __CYGWIN
 #include <netinet/ip6.h>
 #endif
+#include <arpa/inet.h>
 #include <errno.h>
 #include <string.h>
 #include <fcntl.h>
@@ -106,7 +107,7 @@ int parse_play_args (char *buffer, pcap_pkts *pkts)
 void do_sleep (struct timeval *, struct timeval *, struct timeval *, struct timeval *);
 void send_packets_cleanup(void *arg)
 {
-  int sock = (int) arg;
+  int sock = (int) ((long) arg);
 
   // Close send socket
   close(sock);
